requirements.txt
playwright
python-dotenv
Pillow
2captcha-python
